random-mcp/pyproject.toml
[project]
name = "random-mcp"
version = "0.2.2"
description = "Random Mcp project"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "asyncclick>=8.1.8",
    "fastmcp>=2.3.5",
    "gitpython>=3.1.44",
    "mcp>=1.9.0",
]

[tool.poetry]
packages = [
    { include = "random_mcp", from = "src" }
]
[tool.hatch.metadata]
allow-direct-references=true

[project.scripts]
random-mcp = "random_mcp.main:run_mcp"

[tool.pytest.ini_options]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "session"

[dependency-groups]
dev = [
  "pytest",
  "pytest-mock",
  "pytest-asyncio",
  "ruff"
]
lint = [
  "ruff"
]

[build-system]
requires = ["hatchling", "uv-dynamic-versioning>=0.7.0"]
build-backend = "hatchling.build"

[tool.ruff]
extend="../pyproject.toml"
